flask_app/requirements.txt
Flask
jsonschema
psycopg2-binary
